move, seal and rename `AnyBound` to `BoundObject`
